fix(core): reduce time it takes to require nx commands
