+++ packages/nx/src/command-line/generate/command-object.ts
@@ -1,6 +1,4 @@
-import { CommandModule, Argv } from 'yargs';
-import { getCwd } from '../../utils/path';
-import { linkToNxDevAndExamples } from '../yargs-utils/documentation';
+import { Argv, CommandModule } from 'yargs';
 import { withVerbose } from '../yargs-utils/shared-options';
 
 export const yargsGenerateCommand: CommandModule = {
@@ -13,7 +11,7 @@ export const yargsGenerateCommand: CommandModule = {
     // Remove the command from the args
     args._ = args._.slice(1);
 
-    process.exit(await (await import('./generate')).generate(getCwd(), args));
+    process.exit(await (await import('./generate')).generate(args));
   },
 };
