+++ packages/nx/src/command-line/migrate/command-object.ts
@@ -1,15 +1,5 @@
 import { Argv, CommandModule } from 'yargs';
-import * as path from 'path';
-import { runNxSync } from '../../utils/child-process';
 import { linkToNxDevAndExamples } from '../yargs-utils/documentation';
-import { execSync } from 'child_process';
-import {
-  copyPackageManagerConfigurationFiles,
-  detectPackageManager,
-  getPackageManagerCommand,
-} from '../../utils/package-manager';
-import { writeJsonFile } from '../../utils/fileutils';
-import { workspaceRoot } from '../../utils/workspace-root';
 import { withVerbose } from '../yargs-utils/shared-options';
 
 export const yargsMigrateCommand: CommandModule = {
@@ -19,8 +9,8 @@ export const yargsMigrateCommand: CommandModule = {
   - Run migrations (e.g., nx migrate --run-migrations=migrations.json). Use flag --if-exists to run migrations only if the migrations file exists.`,
   builder: (yargs) =>
     linkToNxDevAndExamples(withMigrationOptions(yargs), 'migrate'),
-  handler: () => {
-    runMigration();
+  handler: async () => {
+    (await import('./migrate')).runMigration();
     process.exit(0);
   },
 };
@@ -103,105 +93,4 @@ function withMigrationOptions(yargs: Argv) {
         return true;
       }
     );
-}
-
-function runMigration() {
-  const runLocalMigrate = () => {
-    runNxSync(`_migrate ${process.argv.slice(3).join(' ')}`, {
-      stdio: ['inherit', 'inherit', 'inherit'],
-    });
-  };
-  if (process.env.NX_MIGRATE_USE_LOCAL === undefined) {
-    const p = nxCliPath();
-    if (p === null) {
-      runLocalMigrate();
-    } else {
-      // ensure local registry from process is not interfering with the install
-      // when we start the process from temp folder the local registry would override the custom registry
-      if (
-        process.env.npm_config_registry &&
-        process.env.npm_config_registry.match(
-          /^https:\/\/registry\.(npmjs\.org|yarnpkg\.com)/
-        )
-      ) {
-        delete process.env.npm_config_registry;
-      }
-      execSync(`${p} _migrate ${process.argv.slice(3).join(' ')}`, {
-        stdio: ['inherit', 'inherit', 'inherit'],
-        windowsHide: false,
-      });
-    }
-  } else {
-    runLocalMigrate();
-  }
-}
-
-function nxCliPath() {
-  const version = process.env.NX_MIGRATE_CLI_VERSION || 'latest';
-  try {
-    const packageManager = detectPackageManager();
-    const pmc = getPackageManagerCommand(packageManager);
-
-    const { dirSync } = require('tmp');
-    const tmpDir = dirSync().name;
-    writeJsonFile(path.join(tmpDir, 'package.json'), {
-      dependencies: {
-        nx: version,
-      },
-      license: 'MIT',
-    });
-    copyPackageManagerConfigurationFiles(workspaceRoot, tmpDir);
-    if (pmc.preInstall) {
-      // ensure package.json and repo in tmp folder is set to a proper package manager state
-      execSync(pmc.preInstall, {
-        cwd: tmpDir,
-        stdio: ['ignore', 'ignore', 'ignore'],
-        windowsHide: false,
-      });
-      // if it's berry ensure we set the node_linker to node-modules
-      if (packageManager === 'yarn' && pmc.ciInstall.includes('immutable')) {
-        execSync('yarn config set nodeLinker node-modules', {
-          cwd: tmpDir,
-          stdio: ['ignore', 'ignore', 'ignore'],
-          windowsHide: false,
-        });
-      }
-    }
-
-    execSync(pmc.install, {
-      cwd: tmpDir,
-      stdio: ['ignore', 'ignore', 'ignore'],
-      windowsHide: false,
-    });
-
-    // Set NODE_PATH so that these modules can be used for module resolution
-    addToNodePath(path.join(tmpDir, 'node_modules'));
-    addToNodePath(path.join(workspaceRoot, 'node_modules'));
-
-    return path.join(tmpDir, `node_modules`, '.bin', 'nx');
-  } catch (e) {
-    console.error(
-      `Failed to install the ${version} version of the migration script. Using the current version.`
-    );
-    if (process.env.NX_VERBOSE_LOGGING === 'true') {
-      console.error(e);
-    }
-    return null;
-  }
-}
-
-function addToNodePath(dir: string) {
-  // NODE_PATH is a delimited list of paths.
-  // The delimiter is different for windows.
-  const delimiter = require('os').platform() === 'win32' ? ';' : ':';
-
-  const paths = process.env.NODE_PATH
-    ? process.env.NODE_PATH.split(delimiter)
-    : [];
-
-  // Add the tmp path
-  paths.push(dir);
-
-  // Update the env variable.
-  process.env.NODE_PATH = paths.join(delimiter);
 }
