+++ packages/nx/src/utils/command-line-utils.ts
@@ -6,6 +6,7 @@ import { NxJsonConfiguration } from '../config/nx-json';
 import { execSync } from 'child_process';
 import { ProjectGraph } from '../config/project-graph';
 import { workspaceRoot } from './workspace-root';
+import { readParallelFromArgsAndEnv } from '../command-line/yargs-utils/shared-options';
 
 export interface RawNxArgs extends NxArgs {
   prod?: boolean;
@@ -189,27 +190,6 @@ export function splitArgsIntoNxArgsAndOverrides(
   nxArgs['parallel'] = readParallelFromArgsAndEnv(args);
 
   return { nxArgs, overrides } as any;
-}
-
-export function readParallelFromArgsAndEnv(args: { [k: string]: any }) {
-  if (args['parallel'] === 'false' || args['parallel'] === false) {
-    return 1;
-  } else if (
-    args['parallel'] === 'true' ||
-    args['parallel'] === true ||
-    args['parallel'] === '' ||
-    // dont require passing --parallel if NX_PARALLEL is set, but allow overriding it
-    (process.env.NX_PARALLEL && args['parallel'] === undefined)
-  ) {
-    return Number(
-      args['maxParallel'] ||
-        args['max-parallel'] ||
-        process.env.NX_PARALLEL ||
-        3
-    );
-  } else if (args['parallel'] !== undefined) {
-    return Number(args['parallel']);
-  }
 }
 
 function normalizeNxArgsRunner(
